Remove pwa from withPWA return value
In Next 12.2.3 a change was added saying that the `next.config.js` "must NOT have additional properties" and that the `pwa` property was considered as an additional property. Therefore, this commit deletes the `pwa` property from the calculated next config in the default exported function in index.js (aka the `withPWA` function). The returned next config therefore does not have the `pwa` field and the warning does not appear.